+++ changelog.rst
@@ -18,6 +18,7 @@ CHANGELOG
     + Add `tunnel_bindings` property
     + Several bugfixes (#49, #56, #57, #59, #60, #62, #64, #66, ...)
       (`Pahaz Blinov`_, `JM Fernández`_)
+    + Add TRACE logging level (`JM Fernández`_)
     + Tests refactoring
 - v.0.0.8 (`JM Fernández`_)
     + Merge `#31`_: Support Unix domain socket (local) forwarding (`Dan Harbin`_)